# ENH: Improve `get` performance