--- a/bids/layout/db.py
+++ b/bids/layout/db.py
@@ -79,9 +79,11 @@ class ConnectionManager:
                 'sqlite://',  # In memory database
                 connect_args={'check_same_thread': False},
                 poolclass=StaticPool)
-            
+
         def regexp(expr, item):
             """Regex function for SQLite's REGEXP."""
+            if not isinstance(item, str):
+                return False
             reg = re.compile(expr, re.I)
             return reg.search(item) is not None