Currently, selecting by tags is done using subqueries. For example, using the `get_fieldmap` function leads to the following database query containing three subqueries:
```sql
SELECT files.path, files.filename, files.dirname, files.is_dir, files.class_
FROM files JOIN tags ON files.path = tags.file_path
WHERE files.is_dir = 0 AND (EXISTS (SELECT 1
FROM tags
WHERE files.path = tags.file_path AND tags.entity_name = 'subject' AND (tags._value REGEXP '01'))) AND (EXISTS (SELECT 1
FROM tags
WHERE files.path = tags.file_path AND tags.entity_name = 'suffix' AND (tags._value REGEXP '(phase1|phasediff|epi|fieldmap)'))) AND (EXISTS (SELECT 1
FROM tags
WHERE files.path = tags.file_path AND tags.entity_name = 'extension' AND ((tags._value REGEXP 'nii.gz') OR (tags._value REGEXP 'nii'))))
```

If we use the `EXPLAIN QUERY PLAN` command, we can see that each subquery will be executed once for each entry in the files table, leading to slow performance (this is what `CORRELATED SCALAR SUBQUERY` means, as far as I know).
```
SCAN TABLE tags USING COVERING INDEX sqlite_autoindex_tags_1
SEARCH TABLE files USING INDEX sqlite_autoindex_files_1 (path=?)
CORRELATED SCALAR SUBQUERY 1'
SEARCH TABLE tags USING INDEX sqlite_autoindex_tags_1 (file_path=? AND entity_name=?)
CORRELATED SCALAR SUBQUERY 2
SEARCH TABLE tags USING INDEX sqlite_autoindex_tags_1 (file_path=? AND entity_name=?)
CORRELATED SCALAR SUBQUERY 3
SEARCH TABLE tags USING INDEX sqlite_autoindex_tags_1 (file_path=? AND entity_name=?)
```

I propose to change the `_build_file_query` function to use equi-joins to achieve the same behavior. This way, only one search of the tags table per filter needs to be performed. This idea is based on the following StackOverflow thread <https://stackoverflow.com/questions/30125597/dynamic-columns-in-database-tables-vs-eav/30197015>.
```sql
SELECT files.path, files.filename, files.dirname, files.is_dir, files.class_
FROM files 
LEFT OUTER JOIN tags AS tags_1 ON files.path = tags_1.file_path AND tags_1.entity_name = 'subject' 
LEFT OUTER JOIN tags AS tags_2 ON files.path = tags_2.file_path AND tags_2.entity_name = 'extension' 
LEFT OUTER JOIN tags AS tags_3 ON files.path = tags_3.file_path AND tags_3.entity_name = 'suffix'
WHERE files.is_dir = 0 AND (tags_1._value REGEXP '01') AND ((tags_2._value REGEXP 'nii') OR (tags_2._value REGEXP 'nii.gz')) AND (tags_3._value REGEXP '(phase1|phasediff|epi|fieldmap)') GROUP BY files.path
```
```
SCAN TABLE files USING INDEX sqlite_autoindex_files_1
SEARCH TABLE tags AS tags_1 USING INDEX sqlite_autoindex_tags_1 (file_path=? AND entity_name=?)
SEARCH TABLE tags AS tags_2 USING INDEX sqlite_autoindex_tags_1 (file_path=? AND entity_name=?)
SEARCH TABLE tags AS tags_3 USING INDEX sqlite_autoindex_tags_1 (file_path=? AND entity_name=?)
```

However, I am not sure if there are any edge cases that require the subquery approach.