@effigies 2021-05-12
Hitting the detached session heisenbug... Restarted.

@tyarkoni Do you know of edge cases that might not be covered by tests?