@tyarkoni 2021-05-12
I can't think of any edge cases per se (though there *are* good reasons for the convoluted logic in `_build_file_query`, as discussed below), and in principle, improved performance would be great. That said, a couple of concerns jump to mind:

* Do we actually have practical concerns about `.get()` performance? My sense is that nearly all, if not all, past complaints about performance stem from issues with indexing at layout construction (still very suboptimal), and not with querying once the layout has been built. Are you (or someone else) experiencing noticeable slowdowns from `.get()`, @HippocampusGirl?

* The actual queries in `_build_file_query()` are currently built dynamically with SQLAlchemy. There's no question the ORM layer adds a lot of overhead, but that does keep the code much more readable, and I'm not sure I'd want to throw it out in favor of raw SQL, if that's what's being proposed. Part of the reason querying is inefficient (even for SQLAlchemy) is that we have quite a bit of logic in order to deal with complexity in the input arguments. For example, users can in principle match each entity against either a single value, or against a list. Plus, there are special `Enums` we define—e.g., to indicate that a value for a tag must be set in the DB, but it can be any defined value. I don't doubt one could write corresponding logic to build raw SQL queries, but I worry that that would look much uglier and be far less maintainable, and my inclination at this point is to prioritize code clarity over performance (unless we're talking about very noticeable gains). OTOH, if the idea is to keep using SQLAlchemy, but to do it in an optimized fashion, then that sounds good to me.

Anyway, if you want to work up a sketch of what this would look like, @HippocampusGirl (or even a full PR), we're happy to review it. But whatever solution you propose should probably retain the SQLAlchemy layer unless there are practically important performance gains we're leaving on the table.