@tyarkoni 2021-05-12
Oh, it's also worth noting that the SQLAlchemy API changed in recent builds, and we're currently pinning to an older version to avoid having to do the migration work (which, at a glance, looks non-trivial). I'm not sure how much of those changes are performance-related, but it's conceivable that simply updating to use a recent SQLAlchemy build would improve performance, or at least make it easier to dynamically build more optimized queries. That might be worth looking into as well, and would be much appreciated independently of performance gains, as the longer we delay the migration, the more confusing it will become for people familiar with bleeding-edge SQLAlchemy.