@tyarkoni 2021-05-12
Ignore dumb earlier comments, @HippocampusGirl. I don't read good. Code looks good. Would be nice to see some performance benchmarks to verify queries are faster (I get that they *should* be, but...).

Also, I thought we had successfully killed off the detached session bug (am I misremembering, @effigies?), so if these changes reintroduce it, that's something we'll definitely need to address. Unfortunately I don't remember what either the cause or solution of that bug was, so it might require some digging through the archives.