@effigies 2021-05-12
No, it's been popping up periodically (albeit less frequently than it once did) on `master`.