@effigies 2021-05-12
The old issue was https://github.com/bids-standard/pybids/issues/625.

Not sure if that's actually the same error. Unfortunately, re-running seems to purge old logs, rather than get a whole new run.