@tyarkoni 2021-05-12
> No, it's been popping up periodically (albeit less frequently than it once did) on master.

Oh, in that case, objection withdrawn. If you're confident there are unlikely to be any potential downsides to this, @HippocampusGirl, I'm fine merging as-is. Though if it's not too much of a hassle, I do think it would be nice to have some sense of what kind of performance improvement we're looking at in practice.