@HippocampusGirl 2021-05-13
I realize I probably should have put some more effort into the explanation around my pull request, so I will try my best to take care of that now. We recently converted a large dataset to BIDS format (more than a thousand subjects with more than ten images each) using `heudiconv`. We then wrote a script that calls `layout.get_fieldmap` for every functional image, which took much longer than I thought it would (a few hours). So I decided to investigate :-) 

I discovered your recommendation to disable `index_metadata`, which speeds up indexing by a lot (https://github.com/bids-standard/pybids/issues/609#issuecomment-627529703). However, it also breaks `layout.get_fieldmap`, because now we are missing the `IntendedFor` metadata in the database. 

> Do we actually have practical concerns about .get() performance?

I wouldn't say that it's an urgent concern, but when fMRIPrep creates a workflow, it spends around one third of its time on `get` (see image and https://github.com/nipreps/fmriprep/pull/2406). In any case, the correlated subqueries lead to polynomial time complexity, meaning that the performance impact will be relevant for big datasets and negligible for small ones. So my answer would be that it depends. 
![remmina_mindandbrain server_localhost_2021513-72742](https://user-images.githubusercontent.com/789054/118093913-612e6f00-b3ce-11eb-83e5-b8347eba5be4.png)

> I do think it would be nice to have some sense of what kind of performance improvement we're looking at in practice.

Before
```
%timeit list(map(layout.get_fieldmap, layout.get(datatype="func", task="rest", extension=["nii", "nii.gz"], return_type="filename")[:100]))
53.1 s ± 734 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After
```
%timeit list(map(layout.get_fieldmap, layout.get(datatype="func", task="rest", extension=["nii", "nii.gz"], return_type="filename")[:100]))
21.9 s ± 41.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Note that `%timeit` is probably not the best way to profile this, because queries and results might be cached by `SQLAlchemy` (but I'm not sure about that).