@effigies 2021-05-13
A 60% speedup seems well worth it. Thanks for the metrics.