@HippocampusGirl 2021-05-13
> If you're confident there are unlikely to be any potential downsides to this

I am reasonably confident (based on my limited knowledge of BIDS), and if any issue occurs around this change, I will be there to investigate/fix it (at least until my contract ends late next year). 

I also just added two small adjustments that further increase performance by around 10% over the previous optimizations.
```
%timeit list(map(layout.get_fieldmap, layout.get(datatype="func", task="rest", extension=["nii", "nii.gz"], return_type="filename")[:100]))
17.4 s ± 103 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```