@effigies 2021-05-13
67%! 😍