@tyarkoni 2021-05-13
Also seems to shave off around 8% of the runtime of the entire test suite (at least on my machine), so definitely a meaningful impact. Thanks again!